+++ src/__tests__/checkpoints.test.ts
@@ -10,6 +10,8 @@ import { TypedEmitter } from "../core/events.js";
 import { InMemoryTaskStore, InMemoryRunStore, createTestTask, createMockStores } from "./fixtures.js";
 import type { OrchestratorContext } from "../core/orchestrator-context.js";
 import type { PolpoConfig, Task, Mission, MissionCheckpoint } from "../core/types.js";
+import { FileCheckpointStore } from "../stores/file-checkpoint-store.js";
+import { FileDelayStore } from "../stores/file-delay-store.js";
 
 // ── Helpers ──────────────────────────────────────────
 
@@ -103,7 +105,11 @@ describe("Checkpoints", () => {
 
   beforeEach(async () => {
     tmpDir = mkdtempSync(join(tmpdir(), "polpo-test-cp-"));
-    ctx = createMockCtx({ polpoDir: tmpDir });
+    ctx = createMockCtx({
+      polpoDir: tmpDir,
+      checkpointStore: new FileCheckpointStore(tmpDir),
+      delayStore: new FileDelayStore(tmpDir),
+    });
     taskMgr = new TaskManager(ctx);
     agentMgr = new AgentManager(ctx);
     missionExec = new MissionExecutor(ctx, taskMgr, agentMgr);
